% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_pv.R
\name{fisher_pv}
\alias{fisher_pv}
\title{Fisher randomization test.}
\usage{
fisher_pv(value, group, Tfun, nsim = 1000)
}
\arguments{
\item{value}{Value variable of length \code{nobs}.}

\item{group}{Grouping variable of length \code{nobs}.}

\item{Tfun}{Function taking vectorized arguments \code{value} and \code{group}, returning \code{ntest} test statistics.}

\item{nsim}{Number of randomizations to perform.}
}
\value{
A two column matrix with \code{ntest} rows and columns:
\describe{
\item{\code{Tobs}}{A vector of length \code{ntest}, where \code{Tobs[i]} is the observed value of test statistic \code{i}.}
\item{\code{pval}}{A vector of \code{ntest} p-values of the form \code{Pr(Tsim[i] > Tobs[i])}, where \code{Tobs[i] = Tfun(value, group)[i]} and where \code{Tsim} is computed by randomly reallocating \code{value} to the given vector \code{group}.}
}
}
\description{
Fisher randomization test.
}
