% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aqoyr.R
\name{get_aqoyr}
\alias{get_aqoyr}
\title{Query a full year station/pollutant table from the AQO website.}
\usage{
get_aqoyr(station_id, poll_id, year, start_month, end_month, fill = FALSE)
}
\arguments{
\item{station_id}{Recording station code (integer).}

\item{poll_id}{Pollutant code (integer).}

\item{year}{Year (between 2007 and 2017).}

\item{start_month}{Optional starting month (integer, default = 1).}

\item{end_month}{Optional ending month (integer, default = 12).}

\item{fill}{Whether to fill tables with missing entries with \code{NA}.}
}
\value{
A \verb{365 x 26} table of pollutant concentrations, where each row is a day and the columns are \code{Station} (which is \code{station_id}) \code{Date} (Y-m-d), and hourly time in \code{H01:H23} format.
}
\description{
Query a full year station/pollutant table from the AQO website.
}
\details{
Times are always in EST, i.e., on a clock which does not change back and forth from DST.
}
