% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance.R
\docType{data}
\name{irradiance_data}
\alias{irradiance_data}
\title{Irradiance data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2806 rows and 3 columns.
}
\usage{
irradiance_data
}
\description{
TBD.
}
\keyword{datasets}
