% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance.R
\docType{data}
\name{irradiance_pval}
\alias{irradiance_pval}
\title{Irradiance p-values.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 6 columns.
}
\usage{
irradiance_pval
}
\description{
TBD.
}
\keyword{datasets}
