% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutant.R
\docType{data}
\name{pollutant_data}
\alias{pollutant_data}
\title{Pollutant data from the AIQ website.}
\format{
A data frame with 176,039 rows and columns:
\describe{
\item{Date}{The date at which the concentration was measured.}
\item{Station}{The name of the monitoring station.}
\item{Pollutant}{The name of the pollutant being measured: \code{O3} (ppb), \code{PM25} (micrograms / m^3), \code{NO2} (ppb), \code{SO2} (ppb), \code{CO} (ppm).}
\item{Concentration}{The average daily concentration from hourly measurements.}
}
}
\source{
\url{http://www.airqualityontario.com/}.
}
\usage{
pollutant_data
}
\description{
A dataset containing pollutant concentrations between 2017-2020 at the station/pollutant combinations in \link{pollutant_info}.
}
\keyword{datasets}
