% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutant.R
\docType{data}
\name{pollutant_info}
\alias{pollutant_info}
\title{Pollutant information from the AQO website.}
\format{
A data frame with 180 rows and columns:
\describe{
\item{station}{The name of the monitoring station.}
\item{station_id}{The ID of the monitoring station.}
\item{pollutant}{The type of air pollutant: \code{O3}, \code{PM25}, \code{NO2}, \code{CO}, or \code{SO2}.}
\item{poll_id}{The ID of the air pollutant.}
\item{has_poll}{Whether the AIQ website has any records for the given station/pollutant combination.}
}
}
\source{
\url{http://www.airqualityontario.com/}.
}
\usage{
pollutant_info
}
\description{
A dataset listing the monitoring stations and pollutants from the Air Quality Ontario website.
}
\keyword{datasets}
