% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutant.R
\docType{data}
\name{pollutant_pval}
\alias{pollutant_pval}
\title{Pollutant p-values.}
\format{
A data frame with 2,943 rows and columns:
\describe{
\item{Station}{The name of the monitoring station.}
\item{Pollutant}{The name of the pollutant being measured: \code{O3}, \code{PM25}, \code{NO2}, \code{SO2}, \code{CO}.}
\item{Period}{2017-2019 or 2020.}
\item{Month}{Name of the month.}
\item{Ndays}{Number of days from which p-value is computed.}
\item{Median}{Median daily concentration in the given month.}
\item{Pval}{Randomization test p-value in the given month.}
}
}
\usage{
pollutant_pval
}
\description{
A dataset containing randomization test p-values for the pollutant concentrations contained in \link{pollutant_data}.
}
\details{
For a detailed explanation of the p-value calculation and the fields in this dataset, please see \code{vignettes("aq2020")}.
}
\keyword{datasets}
