% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_boxplot.R
\name{pval_boxplot}
\alias{pval_boxplot}
\title{P-value boxplots for AQ2020.}
\usage{
pval_boxplot(data, pval, main, ylab, pt_size = 2, pv_size = 4, legend = TRUE)
}
\arguments{
\item{data}{Tibble containing columns \code{Year}, \code{Month}, \code{Day}, and \code{Value}.}

\item{pval}{Tibble containing columns \code{Period}, \code{Month}, and \code{Value}.}

\item{main}{Title of plot.}

\item{ylab}{Label for y-axis.}
}
\description{
P-value boxplots for AQ2020.
}
