% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_pv.R
\name{Tbin}
\alias{Tbin}
\title{Two-group test statistics.}
\usage{
Tbin(value, group)
}
\arguments{
\item{value}{Value variable of length \code{nobs}.}

\item{group}{Grouping variable of length \code{nobs}.  Must contain exactly two distinct values.}
}
\value{
A vector of length two, containing the absolute difference between group means and group medians, respectively.
}
\description{
Two-group test statistics.
}
