% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tdiff.R
\name{Tdiff}
\alias{Tdiff}
\title{Two-group absolute-difference statistics.}
\usage{
Tdiff(value, group)
}
\arguments{
\item{value}{Value variable of length \code{nobs}.}

\item{group}{Grouping variable of length \code{nobs}.  Must contain exactly two distinct values.}
}
\value{
A vector of length two, containing the absolute difference between group means and group medians, respectively.
}
\description{
Two-group absolute-difference statistics.
}
