% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\docType{data}
\name{temperature_data}
\alias{temperature_data}
\title{Temperature data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20089 rows and 15 columns.
}
\usage{
temperature_data
}
\description{
TBD.
}
\keyword{datasets}
