% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\docType{data}
\name{temperature_info}
\alias{temperature_info}
\title{Temperature information.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 6 columns.
}
\usage{
temperature_info
}
\description{
TBD.
}
\keyword{datasets}
