% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\docType{data}
\name{temperature_pval}
\alias{temperature_pval}
\title{Temperature p-values.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 336 rows and 6 columns.
}
\usage{
temperature_pval
}
\description{
TBD.
}
\keyword{datasets}
