% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_pv.R
\name{Tmulti}
\alias{Tmulti}
\title{Multi-group test statistics.}
\usage{
Tmulti(value, group)
}
\arguments{
\item{value}{Value variable of length \code{nobs}.}

\item{group}{Grouping variable of length \code{nobs}.  Must contain exactly three or more distinct values.}
}
\value{
A vector of length two, containing the variance of within-group means and within-group medians, respectively.
}
\description{
Multi-group test statistics.
}
