% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trange.R
\name{Trange}
\alias{Trange}
\title{Multi-group range statistics.}
\usage{
Trange(value, group)
}
\arguments{
\item{value}{Value variable of length \code{nobs}.}

\item{group}{Grouping variable of length \code{nobs}.  Must contain two or more distinct values.}
}
\value{
A vector of length two, containing the range (max minus min) of within-group means and within-group medians, respectively.
}
\description{
Multi-group range statistics.
}
