% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tvar.R
\name{Tvar}
\alias{Tvar}
\title{Multi-group variance statistics.}
\usage{
Tvar(value, group)
}
\arguments{
\item{value}{Value variable of length \code{nobs}.}

\item{group}{Grouping variable of length \code{nobs}.  Must contain two or more distinct values.}
}
\value{
A vector of length two, containing the variance of within-group means and within-group medians, respectively.
}
\description{
Multi-group variance statistics.
}
