% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq2020-package.R
\docType{package}
\name{aq2020-package}
\alias{aq2020}
\alias{aq2020-package}
\title{aq2020: Statistical Analysis of Air Quality Data}
\description{
Provides the statistical tools created for the analysis in Al-Abadleh et al (2021) <doi:10.1016/j.jhazmat.2021.125445>, including step-by-step instructions to replicate the p-value calculations and figures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlysy/aq2020}
  \item Report bugs at \url{https://github.com/mlysy/aq2020/issues}
}

}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Authors:
\itemize{
  \item Hind Al-Abadleh
  \item Lucas Neil
  \item Priyesh Patel
  \item Wisam Mohammed
  \item Yara Khalaf
}

}
\keyword{internal}
