% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq2020-package.R
\docType{package}
\name{aq2020-package}
\alias{aq2020}
\alias{aq2020-package}
\title{aq2020: Software to Replicate Analysis of AQ2020 Paper}
\description{
Provides the statistical tools created for the analysis in this paper, including step-by-step instructions to replicate the p-value calculations and figures.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Other contributors:
\itemize{
  \item Hind Al-Abadleh [contributor]
  \item Lucas Neil [contributor]
  \item Priyesh Patel [contributor]
  \item Wisam Mohammed [contributor]
  \item Yara Khalaf [contributor]
}

}
\keyword{internal}
