% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_pv.R
\name{fisher_pv}
\alias{fisher_pv}
\title{Compute the p-value of a Fisher randomization test.}
\usage{
fisher_pv(value, group, Tfun, nsim = 1000, strict = TRUE)
}
\arguments{
\item{value}{Vector of length \code{nobs} corresponding to the values from which the test statistics is computed.  That is, \code{value[i]} is the value corresponding to observation \code{i}.}

\item{group}{Vector of length \code{nobs} indicating to which group each observation belongs.  Group membership is determined from the elements of \code{unique(group)}.}

\item{Tfun}{Function taking vector arguments \code{value} and \code{group}, returning \code{ntest} test statistics.}

\item{nsim}{Number of randomizations to perform.}

\item{strict}{Whether the p-value is "strictly greater" or just "greater than or equal" (see 'Details').}
}
\value{
A two column matrix with \code{ntest} rows and columns:
\describe{
\item{\code{Tobs}}{A vector of length \code{ntest}, where \code{Tobs[t]} is the observed value of test statistic \code{t}.}
\item{\code{pval}}{A vector of \code{ntest} p-values of the form \code{Pr(Tsim[t] > Tobs[t])}, where \code{Tobs[t] = Tfun(value, group)[t]} and \code{Tsim[t]} is computed by randomly reallocating \code{value} to the given vector \code{group}.}
}
}
\description{
Compute the p-value of a Fisher randomization test.
}
\details{
Suppose there is only one test statistic \code{ntest = 1}.  Then the Fisher randomization test calculates\preformatted{Pr(Tsim > Tobs),
}

where \code{Tobs = Tfun(value, group)} is the observed value of the test statistic, and \code{Tsim = Tfun(value[irand], group)}, where \code{irand} is a random permutation of the \code{nobs} observations.  When \code{ntest > 1}, \code{\link[=fisher_pv]{fisher_pv()}} performs the calculation above for each test statistic.

When \code{strict = FALSE} the p-value is \code{Pr(Tsim >= Tobs)}.
}
